backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "es-core-news-sm",
    "fastapi>=0.109.0",
    "langchain>=0.1.16",
    "langchain-community>=0.0.32",
    "langchain-openai>=0.1.1",
    "langchain-postgres>=0.0.3",
    "langgraph>=0.2.0",
    "levenshtein>=0.25.0",
    "numpy<2",
    "openai>=1.12.0",
    "pgvector>=0.2.4",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "spacy>=3.7.2",
    "uvicorn>=0.27.0",
]

[tool.uv.sources]
es-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "httpx>=0.26.0",
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]>=0.4.7",
    "pytest>=8.0.0",
]
